+++ cdp-agentkit-core/python/cdp_agentkit_core/actions/__init__.py
@@ -1,6 +1,7 @@
 from cdp_agentkit_core.actions.cdp_action import CdpAction  # noqa: I001
 
 from cdp_agentkit_core.actions.address_reputation import AddressReputationAction
+from cdp_agentkit_core.actions.deploy_contract import DeployContractAction
 from cdp_agentkit_core.actions.deploy_nft import DeployNftAction
 from cdp_agentkit_core.actions.deploy_token import DeployTokenAction
 from cdp_agentkit_core.actions.get_balance import GetBalanceAction
@@ -43,6 +44,7 @@ __all__ = [
     "AddressReputationAction",
     "DeployNftAction",
     "DeployTokenAction",
+    "DeployContractAction",
     "GetBalanceAction",
     "GetBalanceNftAction",
     "GetWalletDetailsAction",